Add configuration options to WoolMonument
This commit introduces a few configurable fields such as origin and orientation
for the wools monument.

Wools are now sorted by their colour codes, which means the placement location
directly at the origin will be for white wool, and this will expand creating
placements for all the other wool colours until damage 15 is reached. This
expansion's direction will depend on the configured orientation.

Supported orientation values and their offset vectors:
 - NORTH(-1, 0, 0),
 - EAST(0, 0, -1),
 - SOUTH(1, 0, 0),
 - WEST(0, 0, 1),
 - UP(0, 1, 0),
 - DOWN(0, -1, 0),
 - NORTH_EAST(NORTH, EAST),
 - NORTH_WEST(NORTH, WEST),
 - SOUTH_EAST(SOUTH, EAST),
 - SOUTH_WEST(SOUTH, WEST),
 - WEST_NORTH_WEST(WEST, NORTH_WEST),
 - NORTH_NORTH_WEST(NORTH, NORTH_WEST),
 - NORTH_NORTH_EAST(NORTH, NORTH_EAST),
 - EAST_NORTH_EAST(EAST, NORTH_EAST),
 - EAST_SOUTH_EAST(EAST, SOUTH_EAST),
 - SOUTH_SOUTH_EAST(SOUTH, SOUTH_EAST),
 - SOUTH_SOUTH_WEST(SOUTH, SOUTH_WEST),
 - WEST_SOUTH_WEST(WEST, SOUTH_WEST),
 - SELF(0, 0, 0);